requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.36
jinja2==3.1.4
httpx==0.27.0
python-dotenv==1.0.1
thefuzz==0.22.1